Add files via upload
Guys, I've completed the function of loading and generating a dictionary which holds the information of all possible points on the map that lazor may pass, blank and fixed positions of blocks, and map's length and height.
Also I've written a function of generation a lazor path, stored as a list of positions. However, the part which stores the possible positions of next block is still flawed, I'll fix that before Oct.30.