Add files via upload
I've made some changes to my two original function so that they now can work perfectly with Xinru's function.
Also I started the solve_lazor() function.
I've achieved the task of generating a combination of blocks, for example, when there are 3 "A" blocks and 1 "C" block. A list will be generated at the very first of this function as [('A', 'A', 'C', 'A'), ('A', 'A', 'A', 'C'), ('A', 'C', 'A', 'A'), ('C', 'A', 'A', 'A')].

But I still have no idea of solving the lazor cuz there would be three situations:
1. There is no place for next block before you put all of them.
2. All blocks have been putted into some places but some target points are not passed.
3. All target points are hit but there are remaining blocks. ( happens a lot when there are a lot of black blocks)

Anyway, I hope you guys can think about it when you have time. If you have different ideas, remember to put them in a new branch.