Add files via upload
I wrote the solve_lazor function. It can solve some problems now, but not all of them. This is because that when generating possible positions for next block, some positions are blank. I'll try to fix this problem.

But at least the speed is guaranteed. 8 block problems now only requires 0.3s to loop through all situations